requirements.txt
streamlit
opencv-python-headless
pytesseract
Pillow
pdf2image
gdown
supervision
groq
huggingface_hub
ultralytics>=8.3.0
